pyproject.toml
[project]
name = "whatsappcomagent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13,<3.14"
dependencies = [
    "aiofiles>=25.1.0",
    "discord-py>=2.6.4",
    "fastapi>=0.129.0",
    "google-api-python-client>=2.190.0",
    "google-auth-httplib2>=0.3.0",
    "google-auth-oauthlib>=1.2.4",
    "httpx>=0.28.1",
    "langchain-anthropic>=1.3.3",
    "langgraph>=1.0.9",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.22",
    "uvicorn[standard]>=0.41.0",
]
